chore(orga): remove component unit tests for Campaign::Activity::Dashboard
